+++ DesignAutomationConsole/Models/OutputModel.cs
@@ -4,7 +4,7 @@ namespace DesignAutomationConsole.Models
 {
     public class OutputModel
     {
-        public string AddInId { get; set; }
+        public string AddInName { get; set; }
         public string VersionName { get; set; }
         public string VersionBuild { get; set; }
         public DateTime TimeStart { get; set; } = DateTime.UtcNow;