+++ RevitAddin.DA.Tester/Models/OutputModel.cs
@@ -10,7 +10,7 @@ namespace RevitAddin.DA.Tester.Models
 {
     public class OutputModel
     {
-        public string AddInId { get; set; }
+        public string AddInName { get; set; }
         public string VersionName { get; set; }
         public string VersionBuild { get; set; }
         public DateTime TimeStart { get; set; } = DateTime.UtcNow;